Stop using smallint for FKs

+++ examples/postgres/sakila-schema.sql
@@ -159,11 +159,11 @@ CREATE TABLE film (
     title character varying(255) NOT NULL,
     description text,
     release_year year,
-    language_id smallint NOT NULL,
-    original_language_id smallint,
-    rental_duration smallint DEFAULT 3 NOT NULL,
+    language_id integer NOT NULL,
+    original_language_id integer,
+    rental_duration integer DEFAULT 3 NOT NULL,
     rental_rate numeric(4,2) DEFAULT 4.99 NOT NULL,
-    length smallint,
+    length integer,
     replacement_cost numeric(5,2) DEFAULT 19.99 NOT NULL,
     rating mpaa_rating DEFAULT 'G'::mpaa_rating,
     last_update timestamp without time zone DEFAULT now() NOT NULL,
@@ -179,8 +179,8 @@ ALTER TABLE public.film OWNER TO postgres;
 --
 
 CREATE TABLE film_actor (
-    actor_id smallint NOT NULL,
-    film_id smallint NOT NULL,
+    actor_id integer NOT NULL,
+    film_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
 
@@ -192,8 +192,8 @@ ALTER TABLE public.film_actor OWNER TO postgres;
 --
 
 CREATE TABLE film_category (
-    film_id smallint NOT NULL,
-    category_id smallint NOT NULL,
+    film_id integer NOT NULL,
+    category_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
 
@@ -232,7 +232,7 @@ CREATE TABLE address (
     address character varying(50) NOT NULL,
     address2 character varying(50),
     district character varying(20) NOT NULL,
-    city_id smallint NOT NULL,
+    city_id integer NOT NULL,
     postal_code character varying(10),
     phone character varying(20) NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
@@ -261,7 +261,7 @@ ALTER TABLE public.city_city_id_seq OWNER TO postgres;
 CREATE TABLE city (
     city_id integer DEFAULT nextval('city_city_id_seq'::regclass) NOT NULL,
     city character varying(50) NOT NULL,
-    country_id smallint NOT NULL,
+    country_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
 
@@ -313,11 +313,11 @@ ALTER TABLE public.customer_customer_id_seq OWNER TO postgres;
 
 CREATE TABLE customer (
     customer_id integer DEFAULT nextval('customer_customer_id_seq'::regclass) NOT NULL,
-    store_id smallint NOT NULL,
+    store_id integer NOT NULL,
     first_name character varying(45) NOT NULL,
     last_name character varying(45) NOT NULL,
     email character varying(50),
-    address_id smallint NOT NULL,
+    address_id integer NOT NULL,
     activebool boolean DEFAULT true NOT NULL,
     create_date date DEFAULT ('now'::text)::date NOT NULL,
     last_update timestamp without time zone DEFAULT now(),
@@ -366,8 +366,8 @@ ALTER TABLE public.inventory_inventory_id_seq OWNER TO postgres;
 
 CREATE TABLE inventory (
     inventory_id integer DEFAULT nextval('inventory_inventory_id_seq'::regclass) NOT NULL,
-    film_id smallint NOT NULL,
-    store_id smallint NOT NULL,
+    film_id integer NOT NULL,
+    store_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
 
@@ -429,8 +429,8 @@ ALTER TABLE public.payment_payment_id_seq OWNER TO postgres;
 
 CREATE TABLE payment (
     payment_id integer DEFAULT nextval('payment_payment_id_seq'::regclass) NOT NULL,
-    customer_id smallint NOT NULL,
-    staff_id smallint NOT NULL,
+    customer_id integer NOT NULL,
+    staff_id integer NOT NULL,
     rental_id integer NOT NULL,
     amount numeric(5,2) NOT NULL,
     payment_date timestamp without time zone NOT NULL
@@ -460,9 +460,9 @@ CREATE TABLE rental (
     rental_id integer DEFAULT nextval('rental_rental_id_seq'::regclass) NOT NULL,
     rental_date timestamp without time zone NOT NULL,
     inventory_id integer NOT NULL,
-    customer_id smallint NOT NULL,
+    customer_id integer NOT NULL,
     return_date timestamp without time zone,
-    staff_id smallint NOT NULL,
+    staff_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
 
@@ -500,10 +500,10 @@ CREATE TABLE staff (
     staff_id integer DEFAULT nextval('staff_staff_id_seq'::regclass) NOT NULL,
     first_name character varying(45) NOT NULL,
     last_name character varying(45) NOT NULL,
-    address_id smallint NOT NULL,
+    address_id integer NOT NULL,
     reports_to_id integer,
     email character varying(50),
-    store_id smallint NOT NULL,
+    store_id integer NOT NULL,
     active boolean DEFAULT true NOT NULL,
     username character varying(16) NOT NULL,
     password character varying(40),
@@ -533,8 +533,8 @@ ALTER TABLE public.store_store_id_seq OWNER TO postgres;
 
 CREATE TABLE store (
     store_id integer DEFAULT nextval('store_store_id_seq'::regclass) NOT NULL,
-    manager_staff_id smallint NOT NULL,
-    address_id smallint NOT NULL,
+    manager_staff_id integer NOT NULL,
+    address_id integer NOT NULL,
     last_update timestamp without time zone DEFAULT now() NOT NULL
 );
